fixed swagger issue and added the code for merge to main workflow.
