+++ .github/workflows/merge-main.yml
@@ -45,7 +45,7 @@ jobs:
     runs-on: ubuntu-22.04
     strategy:
       matrix:
-        name: [backend, database, frontend, init, backend-java, backend-py]
+        name: [backend, database, frontend, init, backend-java, backend-py, backend-go]
         include:
           - name: backend
             file: backend/openshift.deploy.yml
@@ -71,6 +71,11 @@ jobs:
             overwrite: true
             parameters:
               -p MIN_REPLICAS=1 -p MAX_REPLICAS=2
+          - name: backend-go
+            file: backend-go/openshift.deploy.yml
+            overwrite: true
+            parameters: -p MIN_REPLICAS=1 -p MAX_REPLICAS=2
+              -p PROMOTE_MIGRATION=${{ github.repository }}/database-migrations-go:test
     steps:
       - uses: bcgov-nr/action-deployer-openshift@v1.1.1
         with:
@@ -92,7 +97,7 @@ jobs:
     runs-on: ubuntu-22.04
     strategy:
       matrix:
-        name: [backend, database, frontend, init, backend-py,  backend-java]
+        name: [backend, database, frontend, init, backend-py,  backend-java, backend-go]
         include:
           - name: backend
             file: backend/openshift.deploy.yml
@@ -118,6 +123,12 @@ jobs:
             overwrite: true
             parameters:
               -p MIN_REPLICAS=1 -p MAX_REPLICAS=2
+          - name: backend-go
+            file: backend-go/openshift.deploy.yml
+            overwrite: true
+            parameters: -p MIN_REPLICAS=1 -p MAX_REPLICAS=2
+              -p PROMOTE_MIGRATION=${{ github.repository }}/database-migrations-go:test
+
     steps:
       - uses: bcgov-nr/action-deployer-openshift@v1.1.1
         with:
@@ -140,7 +151,7 @@ jobs:
       packages: write
     strategy:
       matrix:
-        component: [backend, database,database-migrations, frontend, database-migrations-py, backend-py, backend-java]
+        component: [backend, database,database-migrations, frontend, database-migrations-py, backend-py, backend-java, backend-go, database-migrations-go]
     steps:
       - uses: shrink/actions-docker-registry-tag@v3
         with:
