+++ backend-go/src/v1/structs/user-struct.go
@@ -15,3 +15,10 @@ type UserAddress struct {
 	ZipCode string `json:"zip_code"`
 	UserID  uint   `json:"user_id"`
 }
+
+// ResponseHTTP represents response body of this API
+type ResponseHTTP struct {
+	Success bool        `json:"success"`
+	Data    interface{} `json:"data"`
+	Message string      `json:"message"`
+}
